+++ src/Trarizon.Library/History.md
@@ -14,7 +14,7 @@
 - Provide generic comparer to `List.Lookup`
 - Add `SortedList<T>`, `EnumerateDescendants`, `ArrayFiller`, `ArrayTruncation`, `ContiguousLinkedList`, `IArrayAllocator`
 - Add `AutoAllocList` with generic Allocator
-- Add `LazyReverse`
+- Add `LazyReverse`, `LookAhead`
 - Fix bugs on `Memento<T>`, remove generic restriction
 - Support `ReversedSpan`, `Stack.AsSpan()` on all runtimes
 - Support collection expression for `ListDictionary`